refactor(blog,brand): colocate MDX modules and schemas in content
Move MDX glob imports and Zod frontmatter schemas into content/index.ts
for blog and brand, removing separate types and modules files.
