+++ apps/blog/components/blog-post-summary/index.ts
@@ -1,1 +1,5 @@
-export { BlogPostSummary, type BlogPostSummaryProps } from "./blog-post-summary"
+export {
+  type BlogPostAuthor,
+  BlogPostSummary,
+  type BlogPostSummaryProps,
+} from "./blog-post-summary"
