+++ apps/blog/components/blog-posts-list/blog-posts-list.test.tsx
@@ -6,8 +6,10 @@ import {
 } from "@tanstack/react-router"
 import { render, screen } from "@testing-library/react"
 import { describe, expect, it } from "vitest"
-import type { BlogPostFrontmatterWithSlug } from "@/blog/types"
-import { BlogPostsList } from "./blog-posts-list"
+import {
+  type BlogPostFrontmatterWithSlug,
+  BlogPostsList,
+} from "./blog-posts-list"
 
 const samplePosts: BlogPostFrontmatterWithSlug[] = [
   {
