--- apps/blog/post-modules.ts
@@ -1,5 +1,0 @@
-import type { MdxModule } from "@/mdx"
-
-export const postModules = import.meta.glob<MdxModule>("./content/*.mdx", {
-  eager: true,
-})
