+++ apps/blog/routes/post.tsx
@@ -5,8 +5,7 @@ import {
 } from "@tanstack/react-router"
 import { ArrowLeft } from "lucide-react"
 import { BlogPostSummary } from "@/blog/components/blog-post-summary"
-import { postModules } from "@/blog/post-modules"
-import { blogPostFrontmatterSchema } from "@/blog/types"
+import { blogPostFrontmatterSchema, postModules } from "@/blog/content"
 import { ProseArticle } from "@/brand/components/prose-article"
 import config from "@/config"
 import { getMdxModuleBySlug, MDXProvider } from "@/mdx"
