--- apps/brand/legal-modules.ts
@@ -1,8 +1,0 @@
-import type { MdxModule } from "@/mdx"
-
-export const legalModules = import.meta.glob<MdxModule>(
-  "./content/legal/*.mdx",
-  {
-    eager: true,
-  },
-)
