--- apps/brand/types.ts
@@ -1,6 +1,0 @@
-import { z } from "zod"
-
-export const legalDocumentFrontmatterSchema = z.object({
-  title: z.string(),
-  date: z.iso.date(),
-})
